Move withRetry functions to RmmRapidsRetryIterator object
